+++ package.json
@@ -24,6 +24,7 @@
     "next": "12.1.0",
     "react": "17.0.2",
     "react-beautiful-dnd": "^13.1.0",
+    "react-countup": "^6.2.0",
     "react-dom": "17.0.2",
     "react-infinite-scroller": "^1.2.5",
     "react-slick": "^0.28.1",
