+++ pages/about/index.tsx
@@ -1,0 +1,5 @@
+import About from "../../src/components/units/about/about.container";
+
+export default function index() {
+  return <About />;
+}
