+++ src/components/commons/doneGirl/doneGirl.styles.ts
@@ -1,0 +1,2060 @@
+import styled from "@emotion/styled";
+
+export const DoneGirlStyle = styled.div`
+  & #log-in-girl-7-path1_tr {
+    animation: log-in-girl-7-path1_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path1_tr__tr {
+    0% {
+      transform: translate(721.694252px, 768.442875px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(721.694252px, 768.442875px) rotate(8deg);
+    }
+    100% {
+      transform: translate(721.694252px, 768.442875px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path2_tr {
+    animation: log-in-girl-7-path2_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path2_tr__tr {
+    0% {
+      transform: translate(764.614975px, 792.540727px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(764.614975px, 792.540727px) rotate(-4deg);
+    }
+    100% {
+      transform: translate(764.614975px, 792.540727px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path2 {
+    animation: log-in-girl-7-path2_s_do 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-path2_s_do {
+    0% {
+      stroke-dashoffset: 0;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      stroke-dashoffset: 60;
+    }
+    100% {
+      stroke-dashoffset: 0;
+    }
+  }
+  #log-in-girl-7-ellipse1_ts {
+    animation: log-in-girl-7-ellipse1_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse1_ts__ts {
+    0% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    12.500000% {
+      transform: translate(295.100006px, 872.539978px) scale(1, 1);
+    }
+    33.333333% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    59.166667% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    71.666667% {
+      transform: translate(295.100006px, 872.539978px) scale(1, 1);
+    }
+    92.500000% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+    }
+    100% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+    }
+  }
+  #log-in-girl-7-ellipse2_ts {
+    animation: log-in-girl-7-ellipse2_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse2_ts__ts {
+    0% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    12.500000% {
+      transform: translate(295.100006px, 872.539978px) scale(1, 1);
+    }
+    33.333333% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    59.166667% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    71.666667% {
+      transform: translate(295.100006px, 872.539978px) scale(1, 1);
+    }
+    92.500000% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+    }
+    100% {
+      transform: translate(295.100006px, 872.539978px) scale(0.733169, 0.733169);
+    }
+  }
+  #log-in-girl-7-ellipse3_ts {
+    animation: log-in-girl-7-ellipse3_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse3_ts__ts {
+    0% {
+      transform: translate(1269.699951px, 247.059998px)
+        scale(0.576432, 0.576432);
+    }
+    49.166667% {
+      transform: translate(1269.699951px, 247.059998px)
+        scale(0.576432, 0.576432);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    61.666667% {
+      transform: translate(1269.699951px, 247.059998px) scale(0, 0);
+    }
+    82.500000% {
+      transform: translate(1269.699951px, 247.059998px)
+        scale(0.576432, 0.576432);
+    }
+    100% {
+      transform: translate(1269.699951px, 247.059998px)
+        scale(0.576432, 0.576432);
+    }
+  }
+  #log-in-girl-7-ellipse4_ts {
+    animation: log-in-girl-7-ellipse4_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse4_ts__ts {
+    0% {
+      transform: translate(446.983927px, 963.74372px) scale(0.576432, 0.576432);
+    }
+    8.333333% {
+      transform: translate(446.983927px, 963.74372px) scale(0.576432, 0.576432);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    20.833333% {
+      transform: translate(446.983927px, 963.74372px) scale(0, 0);
+    }
+    41.666667% {
+      transform: translate(446.983927px, 963.74372px) scale(0.576432, 0.576432);
+    }
+    100% {
+      transform: translate(446.983927px, 963.74372px) scale(0.576432, 0.576432);
+    }
+  }
+  #log-in-girl-7-ellipse5_ts {
+    animation: log-in-girl-7-ellipse5_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse5_ts__ts {
+    0% {
+      transform: translate(1232.819885px, 582.52002px) scale(0.576432, 0.576432);
+    }
+    9.166667% {
+      transform: translate(1232.819885px, 582.52002px) scale(0.576432, 0.576432);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    27.500000% {
+      transform: translate(1232.819885px, 582.52002px) scale(0, 0);
+    }
+    54.166667% {
+      transform: translate(1232.819885px, 582.52002px) scale(0.576432, 0.576432);
+    }
+    100% {
+      transform: translate(1232.819885px, 582.52002px) scale(0.576432, 0.576432);
+    }
+  }
+  #log-in-girl-7-ellipse6_ts {
+    animation: log-in-girl-7-ellipse6_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse6_ts__ts {
+    0% {
+      transform: translate(120.150002px, 700.400024px) scale(0.576432, 0.576432);
+    }
+    23.333333% {
+      transform: translate(120.150002px, 700.400024px) scale(0.576432, 0.576432);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    41.666667% {
+      transform: translate(120.150002px, 700.400024px) scale(0, 0);
+    }
+    68.333333% {
+      transform: translate(120.150002px, 700.400024px) scale(0.576432, 0.576432);
+    }
+    100% {
+      transform: translate(120.150002px, 700.400024px) scale(0.576432, 0.576432);
+    }
+  }
+  #log-in-girl-7-ellipse7_ts {
+    animation: log-in-girl-7-ellipse7_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse7_ts__ts {
+    0% {
+      transform: translate(1217.279968px, 878.75px) scale(0.371945, 0.371945);
+    }
+    59.166667% {
+      transform: translate(1217.279968px, 878.75px) scale(0.371945, 0.371945);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    77.500000% {
+      transform: translate(1217.279968px, 878.75px) scale(1, 1);
+    }
+    100% {
+      transform: translate(1217.279968px, 878.75px) scale(0.371945, 0.371945);
+    }
+  }
+  #log-in-girl-7-ellipse8_ts {
+    animation: log-in-girl-7-ellipse8_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse8_ts__ts {
+    0% {
+      transform: translate(1217.280029px, 878.75px) scale(0.452927, 0.452927);
+    }
+    59.166667% {
+      transform: translate(1217.280029px, 878.75px) scale(0.452927, 0.452927);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    77.500000% {
+      transform: translate(1217.280029px, 878.75px) scale(1, 1);
+    }
+    100% {
+      transform: translate(1217.280029px, 878.75px) scale(0.452927, 0.452927);
+    }
+  }
+  #log-in-girl-7-ellipse9_ts {
+    animation: log-in-girl-7-ellipse9_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse9_ts__ts {
+    0% {
+      transform: translate(527.990021px, 235.740005px) scale(0.579296, 0.579296);
+    }
+    15% {
+      transform: translate(527.990021px, 235.740005px) scale(0.579296, 0.579296);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    37.500000% {
+      transform: translate(527.990021px, 235.740005px) scale(1, 1);
+    }
+    62.500000% {
+      transform: translate(527.990021px, 235.740005px) scale(0.579296, 0.579296);
+    }
+    100% {
+      transform: translate(527.990021px, 235.740005px) scale(0.579296, 0.579296);
+    }
+  }
+  #log-in-girl-7-ellipse10_ts {
+    animation: log-in-girl-7-ellipse10_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse10_ts__ts {
+    0% {
+      transform: translate(527.98999px, 235.740013px) scale(0.733169, 0.733169);
+    }
+    15% {
+      transform: translate(527.98999px, 235.740013px) scale(0.733169, 0.733169);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    37.500000% {
+      transform: translate(527.98999px, 235.740013px) scale(1, 1);
+    }
+    62.500000% {
+      transform: translate(527.98999px, 235.740013px) scale(0.733169, 0.733169);
+    }
+    100% {
+      transform: translate(527.98999px, 235.740013px) scale(0.733169, 0.733169);
+    }
+  }
+  #log-in-girl-7-path3 {
+    animation-name: log-in-girl-7-path3_s_w, log-in-girl-7-path3_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path3_s_w {
+    0% {
+      stroke-width: 12;
+    }
+    50% {
+      stroke-width: 12;
+    }
+    57.500000% {
+      stroke-width: 6;
+    }
+    61.666667% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path3_s_do {
+    0% {
+      stroke-dashoffset: 5972;
+    }
+    50% {
+      stroke-dashoffset: 5972;
+    }
+    66.666667% {
+      stroke-dashoffset: 600;
+    }
+    100% {
+      stroke-dashoffset: 600;
+    }
+  }
+  #log-in-girl-7-path4 {
+    animation-name: log-in-girl-7-path4_s_w, log-in-girl-7-path4_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path4_s_w {
+    0% {
+      stroke-width: 12;
+    }
+    50% {
+      stroke-width: 12;
+    }
+    57.500000% {
+      stroke-width: 6;
+    }
+    61.666667% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path4_s_do {
+    0% {
+      stroke-dashoffset: 5172;
+    }
+    49.916667% {
+      stroke-dashoffset: 5172;
+    }
+    65.250000% {
+      stroke-dashoffset: 200;
+    }
+    100% {
+      stroke-dashoffset: 200;
+    }
+  }
+  #log-in-girl-7-path5 {
+    animation-name: log-in-girl-7-path5_s_w, log-in-girl-7-path5_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path5_s_w {
+    0% {
+      stroke-width: 12;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      stroke-width: 15;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    7.500000% {
+      stroke-width: 10;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10.833333% {
+      stroke-width: 15;
+    }
+    18.333333% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path5_s_do {
+    0% {
+      stroke-dashoffset: 200;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    18.333333% {
+      stroke-dashoffset: 4063;
+    }
+    100% {
+      stroke-dashoffset: 4063;
+    }
+  }
+  #log-in-girl-7-path6 {
+    animation-name: log-in-girl-7-path6_s_w, log-in-girl-7-path6_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path6_s_w {
+    0% {
+      stroke-width: 12;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      stroke-width: 15;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    7.500000% {
+      stroke-width: 10;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10.833333% {
+      stroke-width: 15;
+    }
+    18.333333% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path6_s_do {
+    0% {
+      stroke-dashoffset: 80;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    17.500000% {
+      stroke-dashoffset: 3823;
+    }
+    100% {
+      stroke-dashoffset: 3823;
+    }
+  }
+  #log-in-girl-7-character_to {
+    animation: log-in-girl-7-character_to__to 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-character_to__to {
+    0% {
+      transform: translate(778.790115px, 814.400955px);
+      animation-timing-function: cubic-bezier(0.42, 0, 1, 1);
+    }
+    25% {
+      transform: translate(768.154989px, 785.033458px);
+      animation-timing-function: cubic-bezier(0, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(763.685205px, 812.298016px);
+      animation-timing-function: cubic-bezier(0.42, 0, 1, 1);
+    }
+    75% {
+      transform: translate(768.154989px, 785.033458px);
+      animation-timing-function: cubic-bezier(0, 0, 0.58, 1);
+    }
+    100% {
+      transform: translate(778.790115px, 814.400955px);
+    }
+  }
+  #log-in-girl-7-character_tr {
+    animation: log-in-girl-7-character_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-character_tr__tr {
+    0% {
+      transform: rotate(0.712854deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: rotate(-1.486471deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.38, 1);
+    }
+    100% {
+      transform: rotate(0.712854deg);
+    }
+  }
+  #log-in-girl-7-legs_to {
+    animation: log-in-girl-7-legs_to__to 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-legs_to__to {
+    0% {
+      transform: translate(756.443124px, 904.469978px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(751.551111px, 893.951668px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      transform: translate(756.443124px, 904.469978px);
+    }
+  }
+  #log-in-girl-7-legs_tr {
+    animation: log-in-girl-7-legs_tr__tr 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-legs_tr__tr {
+    0% {
+      transform: rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: rotate(-6deg);
+    }
+    50% {
+      transform: rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: rotate(-6deg);
+    }
+    100% {
+      transform: rotate(0deg);
+    }
+  }
+  #log-in-girl-7-l-leg_tr {
+    animation: log-in-girl-7-l-leg_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-l-leg_tr__tr {
+    0% {
+      transform: translate(776.325144px, 928.875767px) rotate(5.550321deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(776.325144px, 928.875767px) rotate(10.775423deg);
+    }
+    100% {
+      transform: translate(776.325144px, 928.875767px) rotate(5.550321deg);
+    }
+  }
+  #log-in-girl-7-r-leg_tr {
+    animation: log-in-girl-7-r-leg_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-r-leg_tr__tr {
+    0% {
+      transform: translate(716.71749px, 938.094892px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(716.71749px, 938.094892px) rotate(-8deg);
+    }
+    100% {
+      transform: translate(716.71749px, 938.094892px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-torso_tr {
+    animation: log-in-girl-7-torso_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-torso_tr__tr {
+    0% {
+      transform: translate(739.434442px, 829.237049px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(739.434442px, 829.237049px) rotate(6deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      transform: translate(739.434442px, 829.237049px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-l-arm_to {
+    animation: log-in-girl-7-l-arm_to__to 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-l-arm_to__to {
+    0% {
+      transform: translate(751.929314px, 639.600551px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(757.860839px, 640.237351px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(751.180414px, 639.52015px);
+    }
+    100% {
+      transform: translate(751.929314px, 639.600551px);
+    }
+  }
+  #log-in-girl-7-l-arm_tr {
+    animation: log-in-girl-7-l-arm_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-l-arm_tr__tr {
+    0% {
+      transform: rotate(18.682774deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      transform: rotate(23.009626deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10% {
+      transform: rotate(10.97017deg);
+    }
+    50% {
+      transform: rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: rotate(21.041617deg);
+    }
+    100% {
+      transform: rotate(18.682774deg);
+    }
+  }
+  #log-in-girl-7-l-elbow_tr {
+    animation: log-in-girl-7-l-elbow_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-l-elbow_tr__tr {
+    0% {
+      transform: translate(888.953721px, 639.249612px) rotate(4.55908deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      transform: translate(888.953721px, 639.249612px) rotate(-19.263395deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10% {
+      transform: translate(888.953721px, 639.249612px) rotate(2.018707deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    16.666667% {
+      transform: translate(888.953721px, 639.249612px) rotate(-7.49257deg);
+    }
+    50% {
+      transform: translate(888.953721px, 639.249612px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(888.953721px, 639.249612px) rotate(5.134698deg);
+    }
+    100% {
+      transform: translate(888.953721px, 639.249612px) rotate(4.55908deg);
+    }
+  }
+  #log-in-girl-7-l-hand_tr {
+    animation: log-in-girl-7-l-hand_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-l-hand_tr__tr {
+    0% {
+      transform: translate(951.265381px, 489.201854px) rotate(-39.949059deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      transform: translate(951.265381px, 489.201854px) rotate(19.901476deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10% {
+      transform: translate(951.265381px, 489.201854px) rotate(12.012165deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    27.500000% {
+      transform: translate(951.265381px, 489.201854px) rotate(-11.390477deg);
+    }
+    50% {
+      transform: translate(951.265381px, 489.201854px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(951.265381px, 489.201854px) rotate(-44.992934deg);
+    }
+    100% {
+      transform: translate(951.265381px, 489.201854px) rotate(-39.949059deg);
+    }
+  }
+  #log-in-girl-7-path21_tr {
+    animation: log-in-girl-7-path21_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path21_tr__tr {
+    0% {
+      transform: translate(959.66438px, 440.493236px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      transform: translate(959.66438px, 440.493236px) rotate(10.639906deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    16.666667% {
+      transform: translate(959.66438px, 440.493236px) rotate(10.639906deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    29.166667% {
+      transform: translate(959.66438px, 440.493236px) rotate(-1.731745deg);
+    }
+    100% {
+      transform: translate(959.66438px, 440.493236px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-r-arm_to {
+    animation: log-in-girl-7-r-arm_to__to 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-r-arm_to__to {
+    0% {
+      transform: translate(630.907199px, 664.34433px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(640.151692px, 667.069631px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(630.907199px, 664.34433px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(640.151692px, 667.069631px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      transform: translate(630.907199px, 664.34433px);
+    }
+  }
+  #log-in-girl-7-r-arm_tr {
+    animation: log-in-girl-7-r-arm_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-r-arm_tr__tr {
+    0% {
+      transform: rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: rotate(2.801191deg);
+    }
+    50% {
+      transform: rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: rotate(2.801191deg);
+    }
+    100% {
+      transform: rotate(0deg);
+    }
+  }
+  #log-in-girl-7-r-elbow_tr {
+    animation: log-in-girl-7-r-elbow_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-r-elbow_tr__tr {
+    0% {
+      transform: translate(555.56043px, 811.826273px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(555.56043px, 811.826273px) rotate(16.902344deg);
+    }
+    50% {
+      transform: translate(555.56043px, 811.826273px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(555.56043px, 811.826273px) rotate(16.902344deg);
+    }
+    100% {
+      transform: translate(555.56043px, 811.826273px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-r-hand_tr {
+    animation: log-in-girl-7-r-hand_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-r-hand_tr__tr {
+    0% {
+      transform: translate(667.004429px, 733.795561px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(667.004429px, 733.795561px) rotate(-30.290707deg);
+    }
+    50% {
+      transform: translate(667.004429px, 733.795561px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(667.004429px, 733.795561px) rotate(-30.290707deg);
+    }
+    100% {
+      transform: translate(667.004429px, 733.795561px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path31_tr {
+    animation: log-in-girl-7-path31_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path31_tr__tr {
+    0% {
+      transform: translate(673.815588px, 735.693333px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(673.815588px, 735.693333px) rotate(7.424412deg);
+    }
+    50% {
+      transform: translate(673.815588px, 735.693333px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(673.815588px, 735.693333px) rotate(7.424412deg);
+    }
+    100% {
+      transform: translate(673.815588px, 735.693333px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path33_tr {
+    animation: log-in-girl-7-path33_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path33_tr__tr {
+    0% {
+      transform: translate(665.120039px, 691.85394px) rotate(8.121512deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(665.120039px, 691.85394px) rotate(0deg);
+    }
+    50% {
+      transform: translate(665.120039px, 691.85394px) rotate(8.121512deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(665.120039px, 691.85394px) rotate(0deg);
+    }
+    100% {
+      transform: translate(665.120039px, 691.85394px) rotate(8.121512deg);
+    }
+  }
+  #log-in-girl-7-head_tr {
+    animation: log-in-girl-7-head_tr__tr 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-head_tr__tr {
+    0% {
+      transform: translate(705.871305px, 614.365228px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(705.871305px, 614.365228px) rotate(-8.588272deg);
+    }
+    50% {
+      transform: translate(705.871305px, 614.365228px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(705.871305px, 614.365228px) rotate(-8.588272deg);
+    }
+    100% {
+      transform: translate(705.871305px, 614.365228px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-head2_tr {
+    animation: log-in-girl-7-head2_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-head2_tr__tr {
+    0% {
+      transform: translate(702.441048px, 545.758393px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(702.441048px, 545.758393px) rotate(-5.505255deg);
+    }
+    50% {
+      transform: translate(702.441048px, 545.758393px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(702.441048px, 545.758393px) rotate(-5.505255deg);
+    }
+    100% {
+      transform: translate(702.441048px, 545.758393px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path44_to {
+    animation: log-in-girl-7-path44_to__to 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path44_to__to {
+    0% {
+      transform: translate(719.807821px, 512.540409px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(726.089217px, 520.343549px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(721.195931px, 513.325343px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(726.089217px, 520.343549px);
+    }
+    100% {
+      transform: translate(719.807821px, 512.540409px);
+    }
+  }
+  #log-in-girl-7-path44_tr {
+    animation: log-in-girl-7-path44_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path44_tr__tr {
+    0% {
+      transform: rotate(-13.563008deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: rotate(3.448383deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: rotate(-6.938377deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: rotate(6.991027deg);
+    }
+    100% {
+      transform: rotate(-13.563008deg);
+    }
+  }
+  #log-in-girl-7-hair_tk {
+    animation: log-in-girl-7-hair_tk__tk 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-hair_tk__tk {
+    0% {
+      transform: translate(732.607645px, 458.024355px) skewX(1deg) skewY(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(732.607645px, 458.024355px) skewX(-5deg) skewY(3deg);
+    }
+    50% {
+      transform: translate(732.607645px, 458.024355px) skewX(1deg) skewY(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(732.607645px, 458.024355px) skewX(-5deg) skewY(3deg);
+    }
+    100% {
+      transform: translate(732.607645px, 458.024355px) skewX(1deg) skewY(0deg);
+    }
+  }
+  #log-in-girl-7-hair_ts {
+    animation: log-in-girl-7-hair_ts__ts 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-hair_ts__ts {
+    0% {
+      transform: scale(1, 1);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: scale(1.059419, 0.901016);
+    }
+    50% {
+      transform: scale(1, 1);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: scale(1.059419, 0.901016);
+    }
+    100% {
+      transform: scale(1, 1);
+    }
+  }
+  #log-in-girl-7-path48 {
+    animation: log-in-girl-7-path48__m 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-path48__m {
+    0% {
+      d: path(
+        "M725.170000,436.810000C709.550000,415.920000,626.260000,427.310000,616.170000,463.560000C609.760000,486.560000,604.460000,534.350000,599.350000,550.740000C594.240000,567.130000,544.940000,610.110000,554.070000,669.230000C555.794215,670.411761,577.092540,678.328207,577.218730,678.032809C603.435527,616.662099,650.316634,610.281435,656.455978,546.714785C662.289931,486.310155,660.749387,435.507762,725.170000,436.810000Z"
+      );
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      d: path(
+        "M725.170000,436.810000C709.550000,415.920000,645.010172,422.242070,620.019346,453.731100C601.861719,476.610138,588.920271,501.101131,588.591843,544.813177C588.263414,588.525224,576.117192,614.145694,548.435988,653.777810C549.177800,662.379359,551.583625,667.146292,554.070000,669.230000C589.821526,660.387076,650.316634,610.281435,656.455978,546.714785C662.289931,486.310155,660.749387,435.507762,725.170000,436.810000Z"
+      );
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      d: path(
+        "M725.170000,436.810000C709.550000,415.920000,626.260000,427.310000,616.170000,463.560000C609.760000,486.560000,604.460000,534.350000,599.350000,550.740000C594.240000,567.130000,544.940000,610.110000,554.070000,669.230000C555.794215,670.411761,577.092540,678.328207,577.218730,678.032809C603.435527,616.662099,650.316634,610.281435,656.455978,546.714785C662.289931,486.310155,660.749387,435.507762,725.170000,436.810000Z"
+      );
+    }
+  }
+  #log-in-girl-7-path49 {
+    animation: log-in-girl-7-path49__m 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-path49__m {
+    0% {
+      d: path(
+        "M559.470000,637.750000C529.160000,584.750000,552.860144,561.983824,563.711077,548.950273C583.521966,525.154507,594.432127,515.535062,604.236448,485.449198C622.846008,428.343284,678.374497,419.545433,709.055309,433.020606L706.815309,443.020606C677.012531,420.382152,622.416425,447.949007,612.350000,484.190000C605.947726,507.239369,594.158144,525.798313,570.610000,551.710000C559.269106,564.189188,536.880000,581.520000,561.490000,635.140000Z"
+      );
+    }
+    25% {
+      d: path(
+        "M547.009719,634.240356C543.702713,596.607323,581.126035,582.430222,587.985118,561.189000C599.365150,534.145871,590.152422,511.262333,604.812642,478.793232C626.059209,428.914357,677.887116,418.242156,708.647655,435.265303L706.407655,445.265303C675.384820,422.104194,625.991879,447.387062,611.873537,480.737237C599.394269,507.815132,608.917191,535.377964,595.455243,564.818231C588.131493,587.204591,554.799109,597.743653,551.462944,635.713955Z"
+      );
+    }
+    50% {
+      d: path(
+        "M534.549439,630.730712C558.245426,608.464647,575.215150,588.350291,578.082385,558.901399C581.031560,528.610906,585.872717,506.989603,605.388837,472.137266C629.272411,429.485429,677.399735,416.938879,708.240000,437.510000L706,447.510000C673.757108,423.826235,629.567333,446.825117,611.397073,477.284473C592.840812,508.390894,589.499464,530.431286,586.123711,563.400134C582.817105,595.693666,572.718217,613.967305,541.435888,636.287909Z"
+      );
+    }
+    75% {
+      d: path(
+        "M563.601214,639.915125C554.378843,604.301267,573.568948,580.210416,580.535152,558.025453C589.324931,530.033029,590.152422,511.262333,604.812642,478.793232C626.059209,428.914357,677.887116,418.242156,708.647655,435.265303L706.407655,445.265303C675.384820,422.104194,625.991879,447.387062,611.873537,480.737237C599.394269,507.815132,596.502514,530.417680,588.005277,561.654684C581.755969,584.627992,565.488970,597.675725,568.054438,641.388723Z"
+      );
+    }
+    100% {
+      d: path(
+        "M559.470000,637.750000C529.160000,584.750000,552.860144,561.983824,563.711077,548.950273C583.521966,525.154507,594.432127,515.535062,604.236448,485.449198C622.846008,428.343284,678.374497,419.545433,709.055309,433.020606L706.815309,443.020606C677.012531,420.382152,622.416425,447.949007,612.350000,484.190000C605.947726,507.239369,594.158144,525.798313,570.610000,551.710000C559.269106,564.189188,536.880000,581.520000,561.490000,635.140000Z"
+      );
+    }
+  }
+  #log-in-girl-7-path49_tr {
+    animation: log-in-girl-7-path49_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path49_tr__tr {
+    0% {
+      transform: translate(692.109626px, 433.059743px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(692.109626px, 433.059743px) rotate(5.207028deg);
+    }
+    50% {
+      transform: translate(692.109626px, 433.059743px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(692.109626px, 433.059743px) rotate(5.207028deg);
+    }
+    100% {
+      transform: translate(692.109626px, 433.059743px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path51 {
+    animation: log-in-girl-7-path51__m 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-path51__m {
+    0% {
+      d: path(
+        "M528.590000,659.630000C538.530000,601.040000,580.740000,610.440000,593.590000,589.140000C602.400000,574.570000,596.080000,519.140000,604.180000,496.710000C626.560000,434.620000,699.580000,411.330000,725.180000,436.810000L722.940000,446.810000C708.640000,431.630000,649.320000,429.330000,615.940000,488.480000C604.280000,509.130000,619.370000,560.530000,601.190000,593C586.240000,619.710000,546.430000,603.540000,536.480000,662.130000Z"
+      );
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      d: path(
+        "M567.052559,670.553332C552.615952,613.464940,584.009590,597.818236,599.033627,578.577338C614.162724,560.578392,610.366316,525.596627,615.680430,490.104013C624.935683,428.288791,699.580000,411.330000,725.180000,436.810000L722.940000,446.810000C708.640000,431.630000,649.931454,425.276960,627.764279,485.030447C616.290889,515.957939,630.898690,554.052444,605.202794,584.390110C586.062318,607.652093,565.528097,622.416889,574.993608,670.077395Z"
+      );
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      d: path(
+        "M528.590000,659.630000C538.530000,601.040000,580.740000,610.440000,593.590000,589.140000C602.400000,574.570000,596.080000,519.140000,604.180000,496.710000C626.560000,434.620000,699.580000,411.330000,725.180000,436.810000L722.940000,446.810000C708.640000,431.630000,649.320000,429.330000,615.940000,488.480000C604.280000,509.130000,619.370000,560.530000,601.190000,593C586.240000,619.710000,546.430000,603.540000,536.480000,662.130000Z"
+      );
+    }
+  }
+  #log-in-girl-7-path51_tr {
+    animation: log-in-girl-7-path51_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path51_tr__tr {
+    0% {
+      transform: translate(708.823608px, 435.898052px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(708.823608px, 435.898052px) rotate(5deg);
+    }
+    50% {
+      transform: translate(708.823608px, 435.898052px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(708.823608px, 435.898052px) rotate(6.313811deg);
+    }
+    100% {
+      transform: translate(708.823608px, 435.898052px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-path52_tr {
+    animation: log-in-girl-7-path52_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-path52_tr__tr {
+    0% {
+      transform: translate(741.877389px, 461.448006px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    25% {
+      transform: translate(741.877389px, 461.448006px) rotate(-1.495972deg);
+    }
+    50% {
+      transform: translate(741.877389px, 461.448006px) rotate(0deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(741.877389px, 461.448006px) rotate(-1.495972deg);
+    }
+    100% {
+      transform: translate(741.877389px, 461.448006px) rotate(0deg);
+    }
+  }
+  #log-in-girl-7-belt {
+    animation: log-in-girl-7-belt__m 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-belt__m {
+    0% {
+      d: path(
+        "M776.214400,817.696761C776.214400,817.696761,779.770471,846.295504,779.770471,846.295504C741.860471,851.295504,699.314668,852.556041,669.224668,862.716041L665.796641,832.223556C700.466641,821.993556,742.434400,820.656761,776.214400,817.696761Z"
+      );
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      d: path(
+        "M775.622204,818.504548C775.622204,818.504548,771.687167,838.900602,771.699890,838.927585C733.789890,843.927585,699.314668,852.556041,669.224668,862.716041L665.796641,832.223556C700.466641,821.993556,748.051920,817.788298,775.622204,818.504548Z"
+      );
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      d: path(
+        "M776.214400,817.696761C776.214400,817.696761,779.770471,846.295504,779.770471,846.295504C741.860471,851.295504,699.314668,852.556041,669.224668,862.716041L665.796641,832.223556C700.466641,821.993556,742.434400,820.656761,776.214400,817.696761Z"
+      );
+    }
+  }
+  #log-in-girl-7-belt_to {
+    animation: log-in-girl-7-belt_to__to 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-belt_to__to {
+    0% {
+      transform: translate(737.511421px, 839.512105px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: translate(738.622712px, 835.205669px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      transform: translate(737.511421px, 839.512105px);
+    }
+  }
+  #log-in-girl-7-belt_tk {
+    animation: log-in-girl-7-belt_tk__tk 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-belt_tk__tk {
+    0% {
+      transform: skewX(2deg) skewY(3deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: skewX(-4deg) skewY(5deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      transform: skewX(2deg) skewY(3deg);
+    }
+  }
+  #log-in-girl-7-belt_ts {
+    animation: log-in-girl-7-belt_ts__ts 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-belt_ts__ts {
+    0% {
+      transform: scale(1, 1);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    50% {
+      transform: scale(1.025309, 1);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    100% {
+      transform: scale(1, 1);
+    }
+  }
+  #log-in-girl-7-keyframes_to {
+    animation: log-in-girl-7-keyframes_to__to 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-keyframes_to__to {
+    0% {
+      transform: translate(1200px, 500.000504px);
+    }
+    58.333333% {
+      transform: translate(1200px, 500.000504px);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    75% {
+      transform: translate(550.981957px, 489.549451px);
+    }
+    100% {
+      transform: translate(550.981957px, 489.549451px);
+    }
+  }
+  #log-in-girl-7-keyframes_ts {
+    animation: log-in-girl-7-keyframes_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-keyframes_ts__ts {
+    0% {
+      transform: scale(1, 1);
+    }
+    54.166667% {
+      transform: scale(1, 1);
+      animation-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    }
+    58.333333% {
+      transform: scale(0.534953, 0.534953);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    60.833333% {
+      transform: scale(0.35716, 0.35716);
+    }
+    75% {
+      transform: scale(0.418957, 0.418957);
+    }
+    100% {
+      transform: scale(0.418957, 0.418957);
+    }
+  }
+  #log-in-girl-7-keyframe5_ts {
+    animation: log-in-girl-7-keyframe5_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-keyframe5_ts__ts {
+    0% {
+      transform: translate(1507.409843px, 604.590103px) rotate(45deg)
+        scale(0, 0);
+    }
+    41.666667% {
+      transform: translate(1507.409843px, 604.590103px) rotate(45deg)
+        scale(0, 0);
+      animation-timing-function: cubic-bezier(0.215, 0.61, 0.355, 1);
+    }
+    54.166667% {
+      transform: translate(1507.409843px, 604.590103px) rotate(45deg)
+        scale(1, 1);
+    }
+    100% {
+      transform: translate(1507.409843px, 604.590103px) rotate(45deg)
+        scale(1, 1);
+    }
+  }
+  #log-in-girl-7-rect6 {
+    animation: log-in-girl-7-rect6_c_o 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-rect6_c_o {
+    0% {
+      opacity: 0;
+    }
+    54.166667% {
+      opacity: 0;
+      animation-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    }
+    58.333333% {
+      opacity: 1;
+    }
+    100% {
+      opacity: 1;
+    }
+  }
+  #log-in-girl-7-keyframe8_ts {
+    animation: log-in-girl-7-keyframe8_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-keyframe8_ts__ts {
+    0% {
+      transform: translate(1200.006447px, 429.609368px) rotate(45deg)
+        scale(0, 0);
+    }
+    41.666667% {
+      transform: translate(1200.006447px, 429.609368px) rotate(45deg)
+        scale(0, 0);
+      animation-timing-function: cubic-bezier(0.215, 0.61, 0.355, 1);
+    }
+    54.166667% {
+      transform: translate(1200.006447px, 429.609368px) rotate(45deg)
+        scale(1, 1);
+    }
+    100% {
+      transform: translate(1200.006447px, 429.609368px) rotate(45deg)
+        scale(1, 1);
+    }
+  }
+  #log-in-girl-7-pen_to {
+    animation: log-in-girl-7-pen_to__to 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-pen_to__to {
+    0% {
+      transform: translate(1200px, 500px);
+    }
+    33.333333% {
+      transform: translate(1200px, 500px);
+      animation-timing-function: cubic-bezier(0.25, 1, 0.25, 1);
+    }
+    41.666667% {
+      transform: translate(1199.986164px, 480.121636px);
+      animation-timing-function: cubic-bezier(0.215, 0.61, 0.355, 1);
+    }
+    54.166667% {
+      transform: translate(1199.925383px, 637.212931px);
+      animation-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    }
+    58.333333% {
+      transform: translate(1198.002233px, 690.258932px);
+    }
+    100% {
+      transform: translate(1198.002233px, 690.258932px);
+    }
+  }
+  #log-in-girl-7-pen_tr {
+    animation: log-in-girl-7-pen_tr__tr 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-pen_tr__tr {
+    0% {
+      transform: rotate(360deg);
+    }
+    21.666667% {
+      transform: rotate(360deg);
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    33.333333% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(0deg);
+    }
+  }
+  #log-in-girl-7-pen_ts {
+    animation: log-in-girl-7-pen_ts__ts 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-pen_ts__ts {
+    0% {
+      transform: scale(0.340932, 0.340932);
+    }
+    21.666667% {
+      transform: scale(0.340932, 0.340932);
+    }
+    33.333333% {
+      transform: scale(1.136798, 1.136798);
+      animation-timing-function: cubic-bezier(0.25, 1, 0.25, 1);
+    }
+    41.666667% {
+      transform: scale(0.893001, 0.893001);
+      animation-timing-function: cubic-bezier(0.215, 0.61, 0.355, 1);
+    }
+    54.166667% {
+      transform: scale(1.312147, 1.312147);
+    }
+    100% {
+      transform: scale(1.312147, 1.312147);
+    }
+  }
+  #log-in-girl-7-pen {
+    animation: log-in-girl-7-pen_c_o 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-pen_c_o {
+    0% {
+      opacity: 1;
+    }
+    54.166667% {
+      opacity: 1;
+      animation-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    }
+    58.333333% {
+      opacity: 0;
+    }
+    100% {
+      opacity: 0;
+    }
+  }
+  #log-in-girl-7-ellipse11 {
+    animation: log-in-girl-7-ellipse11_s_do 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-ellipse11_s_do {
+    0% {
+      stroke-dashoffset: 359.62;
+    }
+    21.666667% {
+      stroke-dashoffset: 359.62;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    33.333333% {
+      stroke-dashoffset: 719.24;
+    }
+    100% {
+      stroke-dashoffset: 719.24;
+    }
+  }
+  #log-in-girl-7-path54 {
+    animation: log-in-girl-7-path54_s_do 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-path54_s_do {
+    0% {
+      stroke-dashoffset: 120.51;
+    }
+    29.166667% {
+      stroke-dashoffset: 120.51;
+      animation-timing-function: cubic-bezier(0, 0, 0.58, 1);
+    }
+    33.333333% {
+      stroke-dashoffset: 241.02;
+    }
+    100% {
+      stroke-dashoffset: 241.02;
+    }
+  }
+  #log-in-girl-7-triangle_to {
+    animation: log-in-girl-7-triangle_to__to 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-triangle_to__to {
+    0% {
+      transform: translate(1200px, 500px);
+    }
+    21.666667% {
+      transform: translate(1200px, 500px);
+    }
+    70% {
+      transform: translate(1332.262482px, 501.160233px);
+      animation-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    }
+    79.166667% {
+      transform: translate(1200px, 500px);
+    }
+    100% {
+      transform: translate(1200px, 500px);
+    }
+  }
+  #log-in-girl-7-triangle_ts {
+    animation: log-in-girl-7-triangle_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-triangle_ts__ts {
+    0% {
+      transform: rotate(90deg) scale(1, 1);
+    }
+    8.333333% {
+      transform: rotate(90deg) scale(1, 1);
+      animation-timing-function: cubic-bezier(0.42, 0, 1, 1);
+    }
+    10% {
+      transform: rotate(90deg) scale(0.615783, 0.615783);
+      animation-timing-function: cubic-bezier(0, 0, 0.58, 1);
+    }
+    13.333333% {
+      transform: rotate(90deg) scale(1, 1);
+      animation-timing-function: cubic-bezier(0.26, 0.46, 0.995, 1);
+    }
+    15% {
+      transform: rotate(90deg) scale(0, 0);
+      animation-timing-function: cubic-bezier(0, 0, 0.255, 1);
+    }
+    21.666667% {
+      transform: rotate(90deg) scale(0, 0);
+    }
+    70% {
+      transform: rotate(90deg) scale(0, 0);
+      animation-timing-function: cubic-bezier(0.175, 0.885, 0.32, 1.275);
+    }
+    79.166667% {
+      transform: rotate(90deg) scale(1, 1);
+    }
+    100% {
+      transform: rotate(90deg) scale(1, 1);
+    }
+  }
+  #log-in-girl-7-stroke-path_tr {
+    animation: log-in-girl-7-stroke-path_tr__tr 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-stroke-path_tr__tr {
+    0% {
+      transform: translate(1200px, 500px) rotate(270deg);
+    }
+    15% {
+      transform: translate(1200px, 500px) rotate(270deg);
+      animation-timing-function: cubic-bezier(0.99, -0.105, 0.58, 1);
+    }
+    21.666667% {
+      transform: translate(1200px, 500px) rotate(-90deg);
+      animation-timing-function: cubic-bezier(0.44, 0.015, 0.46, 0.99);
+    }
+    33.333333% {
+      transform: translate(1200px, 500px) rotate(-450deg);
+    }
+    100% {
+      transform: translate(1200px, 500px) rotate(-450deg);
+    }
+  }
+  #log-in-girl-7-stroke-path_ts {
+    animation: log-in-girl-7-stroke-path_ts__ts 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-stroke-path_ts__ts {
+    0% {
+      transform: scale(2.07539, 2.07539);
+    }
+    13.333333% {
+      transform: scale(2.07539, 2.07539);
+      animation-timing-function: cubic-bezier(0.26, 0.46, 0.995, 1);
+    }
+    15% {
+      transform: scale(0.1, 0.1);
+      animation-timing-function: cubic-bezier(0, 0, 0.255, 1);
+    }
+    16.666667% {
+      transform: scale(1, 1);
+    }
+    21.666667% {
+      transform: scale(1, 1);
+      animation-timing-function: cubic-bezier(0.44, 0.015, 0.46, 0.99);
+    }
+    33.333333% {
+      transform: scale(2.07539, 2.07539);
+    }
+    100% {
+      transform: scale(2.07539, 2.07539);
+    }
+  }
+  #log-in-girl-7-stroke-path {
+    animation-name: log-in-girl-7-stroke-path_s_w,
+      log-in-girl-7-stroke-path_s_da;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-stroke-path_s_w {
+    0% {
+      stroke-width: 4;
+    }
+    13.333333% {
+      stroke-width: 4;
+      animation-timing-function: cubic-bezier(0.26, 0.46, 0.995, 1);
+    }
+    15% {
+      stroke-width: 110;
+      animation-timing-function: cubic-bezier(0, 0, 0.255, 1);
+    }
+    21.666667% {
+      stroke-width: 110;
+      animation-timing-function: cubic-bezier(0, 0, 0.255, 1);
+    }
+    33.333333% {
+      stroke-width: 4;
+    }
+    100% {
+      stroke-width: 4;
+    }
+  }
+  @keyframes log-in-girl-7-stroke-path_s_da {
+    0% {
+      stroke-dasharray: 345.58, 287.9833;
+    }
+    15% {
+      stroke-dasharray: 345.58, 287.9833;
+      animation-timing-function: cubic-bezier(0, 0, 0.255, 1);
+    }
+    16.666667% {
+      stroke-dasharray: 57.5966, 287.9833;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    18.333333% {
+      stroke-dasharray: 57.5966, 287.9833;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    20% {
+      stroke-dasharray: 110, 287.9833;
+    }
+    21.666667% {
+      stroke-dasharray: 57.5966, 287.9833;
+      animation-timing-function: cubic-bezier(0.44, 0.015, 0.46, 0.99);
+    }
+    33.333333% {
+      stroke-dasharray: 345.58, 287.9833;
+    }
+    100% {
+      stroke-dasharray: 345.58, 287.9833;
+    }
+  }
+  #log-in-girl-7-stroke2 {
+    animation: log-in-girl-7-stroke2_c_o 12000ms linear infinite normal forwards;
+  }
+  @keyframes log-in-girl-7-stroke2_c_o {
+    0% {
+      opacity: 1;
+    }
+    6.666667% {
+      opacity: 1;
+    }
+    7.166667% {
+      opacity: 0;
+    }
+    9.166667% {
+      opacity: 0;
+    }
+    10% {
+      opacity: 1;
+    }
+    11.750000% {
+      opacity: 1;
+    }
+    11.916667% {
+      opacity: 0;
+    }
+    100% {
+      opacity: 0;
+    }
+  }
+  #log-in-girl-7-path55 {
+    animation-name: log-in-girl-7-path55_s_w, log-in-girl-7-path55_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path55_s_w {
+    0% {
+      stroke-width: 12;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      stroke-width: 15;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    7.500000% {
+      stroke-width: 10;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10.833333% {
+      stroke-width: 15;
+    }
+    18.333333% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path55_s_do {
+    0% {
+      stroke-dashoffset: 200;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    18.333333% {
+      stroke-dashoffset: 4063;
+    }
+    100% {
+      stroke-dashoffset: 4063;
+    }
+  }
+  #log-in-girl-7-path56 {
+    animation-name: log-in-girl-7-path56_s_w, log-in-girl-7-path56_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path56_s_w {
+    0% {
+      stroke-width: 12;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    4.166667% {
+      stroke-width: 15;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    7.500000% {
+      stroke-width: 10;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    10.833333% {
+      stroke-width: 15;
+    }
+    18.333333% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path56_s_do {
+    0% {
+      stroke-dashoffset: 80;
+      animation-timing-function: cubic-bezier(0.42, 0, 0.58, 1);
+    }
+    17.500000% {
+      stroke-dashoffset: 3823;
+    }
+    100% {
+      stroke-dashoffset: 3823;
+    }
+  }
+  #log-in-girl-7-path57 {
+    animation-name: log-in-girl-7-path57_s_w, log-in-girl-7-path57_s_do,
+      log-in-girl-7-path57_s_da;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path57_s_w {
+    0% {
+      stroke-width: 6;
+    }
+    26.666667% {
+      stroke-width: 6;
+    }
+    30.833333% {
+      stroke-width: 12;
+    }
+    35% {
+      stroke-width: 6;
+    }
+    100% {
+      stroke-width: 6;
+    }
+  }
+  @keyframes log-in-girl-7-path57_s_do {
+    0% {
+      stroke-dashoffset: 2843;
+    }
+    25% {
+      stroke-dashoffset: 2843;
+    }
+    32.500000% {
+      stroke-dashoffset: 1483.76;
+    }
+    38.333333% {
+      stroke-dashoffset: 104;
+    }
+    100% {
+      stroke-dashoffset: 104;
+    }
+  }
+  @keyframes log-in-girl-7-path57_s_da {
+    0% {
+      stroke-dasharray: 120, 2603;
+    }
+    25% {
+      stroke-dasharray: 120, 2603;
+    }
+    32.500000% {
+      stroke-dasharray: 400, 2603;
+    }
+    39.166667% {
+      stroke-dasharray: 104, 2603;
+    }
+    100% {
+      stroke-dasharray: 104, 2603;
+    }
+  }
+  #log-in-girl-7-path58 {
+    animation-name: log-in-girl-7-path58_s_w, log-in-girl-7-path58_s_do,
+      log-in-girl-7-path58_s_da;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path58_s_w {
+    0% {
+      stroke-width: 6;
+    }
+    26.666667% {
+      stroke-width: 6;
+    }
+    30.833333% {
+      stroke-width: 12;
+    }
+    35% {
+      stroke-width: 6;
+    }
+    100% {
+      stroke-width: 6;
+    }
+  }
+  @keyframes log-in-girl-7-path58_s_do {
+    0% {
+      stroke-dashoffset: 2663;
+    }
+    25% {
+      stroke-dashoffset: 2663;
+    }
+    37.250000% {
+      stroke-dashoffset: 30;
+    }
+    100% {
+      stroke-dashoffset: 30;
+    }
+  }
+  @keyframes log-in-girl-7-path58_s_da {
+    0% {
+      stroke-dasharray: 30, 2603;
+    }
+    25% {
+      stroke-dasharray: 30, 2603;
+    }
+    32.500000% {
+      stroke-dasharray: 200, 2603;
+    }
+    37.916667% {
+      stroke-dasharray: 30, 2603;
+    }
+    100% {
+      stroke-dasharray: 30, 2603;
+    }
+  }
+  #log-in-girl-7-stroke-32 {
+    animation: log-in-girl-7-stroke-32_c_o 12000ms linear infinite normal
+      forwards;
+  }
+  @keyframes log-in-girl-7-stroke-32_c_o {
+    0% {
+      opacity: 1;
+    }
+    56.250000% {
+      opacity: 1;
+    }
+    56.416667% {
+      opacity: 0;
+    }
+    60.750000% {
+      opacity: 0;
+    }
+    61.083333% {
+      opacity: 1;
+    }
+    100% {
+      opacity: 1;
+    }
+  }
+  #log-in-girl-7-path59 {
+    animation-name: log-in-girl-7-path59_s_w, log-in-girl-7-path59_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path59_s_w {
+    0% {
+      stroke-width: 12;
+    }
+    50% {
+      stroke-width: 12;
+    }
+    57.500000% {
+      stroke-width: 6;
+    }
+    61.666667% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path59_s_do {
+    0% {
+      stroke-dashoffset: 5972;
+    }
+    50% {
+      stroke-dashoffset: 5972;
+    }
+    66.666667% {
+      stroke-dashoffset: 600;
+    }
+    100% {
+      stroke-dashoffset: 600;
+    }
+  }
+  #log-in-girl-7-path60 {
+    animation-name: log-in-girl-7-path60_s_w, log-in-girl-7-path60_s_do;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path60_s_w {
+    0% {
+      stroke-width: 12;
+    }
+    50% {
+      stroke-width: 12;
+    }
+    57.500000% {
+      stroke-width: 6;
+    }
+    61.666667% {
+      stroke-width: 12;
+    }
+    100% {
+      stroke-width: 12;
+    }
+  }
+  @keyframes log-in-girl-7-path60_s_do {
+    0% {
+      stroke-dashoffset: 5172;
+    }
+    49.916667% {
+      stroke-dashoffset: 5172;
+    }
+    65.250000% {
+      stroke-dashoffset: 200;
+    }
+    100% {
+      stroke-dashoffset: 200;
+    }
+  }
+  #log-in-girl-7-path61 {
+    animation-name: log-in-girl-7-path61_s_w, log-in-girl-7-path61_s_do,
+      log-in-girl-7-path61_s_da;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path61_s_w {
+    0% {
+      stroke-width: 6;
+    }
+    76.666667% {
+      stroke-width: 6;
+    }
+    80% {
+      stroke-width: 12;
+    }
+    85% {
+      stroke-width: 6;
+    }
+    100% {
+      stroke-width: 6;
+    }
+  }
+  @keyframes log-in-girl-7-path61_s_do {
+    0% {
+      stroke-dashoffset: 2843;
+    }
+    75% {
+      stroke-dashoffset: 2843;
+    }
+    82.500000% {
+      stroke-dashoffset: 1483.76;
+    }
+    88.333333% {
+      stroke-dashoffset: 104;
+    }
+    100% {
+      stroke-dashoffset: 104;
+    }
+  }
+  @keyframes log-in-girl-7-path61_s_da {
+    0% {
+      stroke-dasharray: 120, 2603;
+    }
+    75% {
+      stroke-dasharray: 120, 2603;
+    }
+    82.500000% {
+      stroke-dasharray: 400, 2603;
+    }
+    89.166667% {
+      stroke-dasharray: 104, 2603;
+    }
+    100% {
+      stroke-dasharray: 104, 2603;
+    }
+  }
+  #log-in-girl-7-path62 {
+    animation-name: log-in-girl-7-path62_s_w, log-in-girl-7-path62_s_do,
+      log-in-girl-7-path62_s_da;
+    animation-duration: 12000ms;
+    animation-fill-mode: forwards;
+    animation-timing-function: linear;
+    animation-direction: normal;
+    animation-iteration-count: infinite;
+  }
+  @keyframes log-in-girl-7-path62_s_w {
+    0% {
+      stroke-width: 6;
+    }
+    76.666667% {
+      stroke-width: 6;
+    }
+    80% {
+      stroke-width: 12;
+    }
+    85% {
+      stroke-width: 6;
+    }
+    100% {
+      stroke-width: 6;
+    }
+  }
+  @keyframes log-in-girl-7-path62_s_do {
+    0% {
+      stroke-dashoffset: 2663;
+    }
+    75% {
+      stroke-dashoffset: 2663;
+    }
+    87.250000% {
+      stroke-dashoffset: 30;
+    }
+    100% {
+      stroke-dashoffset: 30;
+    }
+  }
+  @keyframes log-in-girl-7-path62_s_da {
+    0% {
+      stroke-dasharray: 30, 2603;
+    }
+    75% {
+      stroke-dasharray: 30, 2603;
+    }
+    82.500000% {
+      stroke-dasharray: 200, 2603;
+    }
+    87.916667% {
+      stroke-dasharray: 30, 2603;
+    }
+    100% {
+      stroke-dasharray: 30, 2603;
+    }
+  }
+`;
