+++ src/components/units/main/main.presenter.tsx
@@ -13,6 +13,7 @@ export default function MainPage(props: IMainPage) {
   const { data } = useFetchUser();
   return (
     <S.MainBox>
+      <S.MainRenderingBackground>TEAMPLAY</S.MainRenderingBackground>
       <MainSlide />
       <Wrapper paddingTop={0}>
         <S.MainTop>
