+++ src/components/units/main/main.styles.ts
@@ -1,3 +1,5 @@
+import { breakPoints } from "./../../../commons/styles/breakpoint";
+import { keyframes } from "@emotion/react";
 import styled from "@emotion/styled";
 
 export const MainBox = styled.div`
@@ -19,5 +21,39 @@ export const ContentsHeader = styled.div`
     font-size: 20px;
     display: flex;
     align-items: center;
+  }
+`;
+
+const mainIntroAni = keyframes`
+  from{
+    opacity: 1;
+    visibility: visible;
+  }
+  to{
+    opacity: 0;
+    visibility: hidden;
+  }
+`;
+
+export const MainRenderingBackground = styled.div`
+  width: 100%;
+  height: 100vh;
+  background-color: #3894ff;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 700;
+  font-size: 100px;
+  letter-spacing: -1;
+  color: #fff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  animation: ${mainIntroAni} 1s ease-in-out 1s 1 forwards;
+  @media ${breakPoints.mobile} {
+    font-size: 50px;
+  }
+  @media ${breakPoints.tablet} {
+    font-size: 70px;
   }
 `;
