+++ src/components/units/project/complete/complete.container.tsx
@@ -1,0 +1,5 @@
+import ProjectCompletePaymentUI from "./complete.presenter";
+
+export default function ProjectCompletePayment() {
+  return <ProjectCompletePaymentUI />;
+}
